==> EcoType_Forest_Cover/src/step01_project_overview.py <==
# src/step01_project_overview.py
print("="*80)
print("EcoType: Forest Cover Type Prediction Using Machine Learning")
print("""
Problem Type : Multi-Class Classification
Domain       : Environmental & Geospatial ML
Target       : Cover_Type (7 Classes)
""")
print("="*80)
